shared/requirements.txt
# Shared dependencies for ONG Management System models
psycopg2-binary==2.9.9
python-dotenv==1.0.0